Organize src into two folder core and updatees
