+++ src/core/script.js
